bars float

+++ src/clock.rs
@@ -7,14 +7,14 @@ pub(crate) struct ClockState {
 pub struct Clock {
     id: usize,
     bpm: f32,
-    bars: usize,
+    bars: f32,
 }
 
 pub fn clock(id: usize) -> Clock {
     Clock {
         id,
         bpm: 120.0,
-        bars: 1,
+        bars: 1.,
     }
 }
 
@@ -24,7 +24,7 @@ impl Clock {
         self
     }
 
-    pub fn bars(mut self, bars: usize) -> Self {
+    pub fn bars(mut self, bars: f32) -> Self {
         self.bars = bars;
         self
     }
@@ -35,7 +35,7 @@ impl Clock {
 
         let beats_per_second = self.bpm / 60.0;
         let samples_per_beat = sample_rate as f32 / beats_per_second;
-        let samples_per_bar = samples_per_beat * 4.0 * self.bars as f32;
+        let samples_per_bar = samples_per_beat * 4.0 * self.bars;
 
         let position_in_bar = (state.position as f32) % samples_per_bar;
         state.position += 1;
